requirements-uv.txt
python-dotenv
google-api-python-client
youtube-transcript-api
psycopg2-binary
openai
pandas
tqdm
